backend/requirements.txt
requests
flask
python-dotenv
Gunicorn